refactor(codegen): elimina func.rs, move compile_program e helpers
func.rs (686 LOC) particionado em 3 modulos finais:
- compile/program.rs (560 LOC) — compile_program + UserFn struct
  + declare_user_fn + user_symbol_name + fn_signature
- compile/util.rs (130 LOC) — is_lifted_callback + user_call_conv
  + collect_var_decls (compartilhados por user_fn/main_fn/program)

func.rs deletado. lower/mod.rs re-exporta compile::program::compile_program
preservando interface publica (lower::compile_program).

Callers externos (expressions/calls.rs, expressions/mod.rs) atualizados
de lower::func::class_*_name -> lower::compile::class::class_*_name.

Estrutura final do refator (de 9603 LOC iniciais):
  lower/
    analysis/  (4 mods)
    passes/    (8 mods)
    compile/   (6 mods: class, main_fn, mir_route, program, user_fn, util)

Suite verde (11/11 crates).
