+++ crates/rts-codegen/src/codegen/lower/compile/mod.rs
@@ -7,4 +7,6 @@
 pub mod class;
 pub mod main_fn;
 pub mod mir_route;
+pub mod program;
 pub mod user_fn;
+pub mod util;
