+++ crates/rts-codegen/src/codegen/lower/expressions/calls.rs
@@ -14,7 +14,9 @@ use super::members::{
 };
 use super::operators::to_f64;
 use crate::codegen::lower::ctx::{FnCtx, TypedVal, ValTy};
-use crate::codegen::lower::func::{class_getter_name, class_setter_name, class_static_method_name};
+use crate::codegen::lower::compile::class::{
+    class_getter_name, class_setter_name, class_static_method_name,
+};
 
 pub(super) fn lower_call(ctx: &mut FnCtx, call: &CallExpr) -> Result<TypedVal> {
     if matches!(&call.callee, Callee::Super(_)) {
